backend/requirements.txt
flask
flask-cors
openai-whisper
pydub
soundfile
openai
python-dotenv
